chore(runtime,core): move CallTargetInterface to core
